requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-wtf==1.1.1
python-dotenv==1.0.0
pymongo==4.5.0
flask-pymongo==2.3.0
openai==1.3.0
azure-identity==1.13.0
pandas
numpy
python-dateutil==2.8.2
pydantic==2.0.3